+++ CHANGELOG.md
@@ -8,6 +8,8 @@ changelog does not include internal changes that do not affect the user.
 
 ## [Unreleased]
 
+## [0.15.0] - 2026-06-15
+
 ### Added
 
 - Added `SDMGradWeighting` from